+++ src/Contour.js
@@ -1,0 +1,8 @@
+export class Contour {
+
+  constructor(contourId, coords) {
+    this.id = contourId
+    this.rawCoords = coords
+  }
+
+}