+++ src/Edge.js
@@ -1,11 +1,12 @@
 export class Edge {
 
-  constructor(p1, p2, edgeType, index) {
+  constructor(p1, p2, edgeType, index, contourId) {
     this.p1 = p1
     this.p2 = p2
     this.edgeType = edgeType
     this.originalIndex = index
 
+    this.polygonContourId = contourId
     this.interiorRing = false
 
     this.minX = Math.min(p1.p[0], p2.p[0])